moving results from Geometry to Analytical
