+++ Analytical_oM/Results/Geometry/Enums/MeshResultSmoothingType.cs
