+++ Analytical_oM/Results/Geometry/IMeshNodeResult.cs
