+++ Analytical_oM/Results/Geometry/IMeshResult.cs
