+++ Analytical_oM/Results/Geometry/MeshFaceSolidResult.cs
