+++ Analytical_oM/Results/Geometry/MeshResult.cs
