+++ Analytical_oM/Results/Geometry/SpatioPeriodicalMeshFaceResult.cs
@@ -35,6 +35,11 @@ namespace BH.oM.Geometry.Results
     public class SpatioPeriodicalMeshFaceResult : SpatioPeriodicalResult<IMeshFaceResult>, IMeshFaceResult
     {
         public IAdapterId FaceId { get; set; }
+
+        public int CompareTo(IResult other)
+        {
+            throw new NotImplementedException();
+        }
     }
 }
